AYS-523 | `institution.util.exception` Package Has Been Moved to `institution` Package
