+++ src/main/java/org/ays/auth/service/impl/AdminRegistrationApplicationServiceImpl.java
@@ -19,8 +19,8 @@ import org.ays.auth.util.exception.AysAdminRegistrationApplicationAlreadyRejecte
 import org.ays.auth.util.exception.AysAdminRegistrationApplicationInCompleteException;
 import org.ays.common.model.AysPage;
 import org.ays.common.model.AysPageable;
+import org.ays.institution.exception.AysInstitutionNotExistException;
 import org.ays.institution.port.InstitutionReadPort;
-import org.ays.institution.util.exception.AysInstitutionNotExistException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
